refactor: signin with desktop
