+++ ui/portal/web/rsbuild.config.ts
@@ -40,6 +40,7 @@ export default defineConfig({
           org: process.env.SENTRY_ORG,
           project: process.env.SENTRY_PROJECT,
           authToken: process.env.SENTRY_AUTH_TOKEN,
+          telemetry: false,
           sourcemaps: {
             filesToDeleteAfterUpload: ["build/**/*.map"],
           },
