+++ ui/portal/web/src/components/create-account/CreateAccountDepositStep.tsx
@@ -9,7 +9,7 @@ import type { AccountTypes } from "@left-curve/dango/types";
 import type React from "react";
 
 import { useApp } from "~/hooks/useApp";
-import { Modals } from "../foundation/Modal";
+import { Modals } from "../foundation/RootModal";
 import { toast } from "../foundation/Toast";
 
 export const CreateAccountDepositStep: React.FC = () => {
