+++ ui/portal/web/src/pages/(auth)/_auth.signin.lazy.tsx
@@ -1,20 +1,27 @@
 import { createLazyFileRoute, useSearch } from "@tanstack/react-router";
 
 import { WizardProvider } from "@left-curve/applets-kit";
+import { useEffect } from "react";
 import { Signin } from "~/components/auth/Signin";
+import { Modals } from "~/components/foundation/RootModal";
+import { useApp } from "~/hooks/useApp";
 
 export const Route = createLazyFileRoute("/(auth)/_auth/signin")({
   component: SigninComponent,
 });
 
 function SigninComponent() {
+  const { showModal } = useApp();
   const { socketId } = useSearch({ strict: false });
 
+  useEffect(() => {
+    if (socketId) showModal(Modals.SignWithDesktop, { socketId });
+  }, []);
+
   return (
-    <WizardProvider wrapper={<Signin />} defaultData={{ socketId }} startIndex={socketId ? 2 : 0}>
+    <WizardProvider wrapper={<Signin />}>
       <Signin.Username />
       <Signin.Credential />
-      <Signin.Mobile />
     </WizardProvider>
   );
 }
